requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
google-cloud-bigquery==3.12.0
db-dtypes==1.2.0
pandas==2.2.0
jinja2==3.1.2
gunicorn==21.2.0
#gspread==6.1.2
gspread==6.2.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-multipart==0.0.9